+++ server/odc-service/src/main/java/com/oceanbase/odc/service/connection/logicaldatabase/core/executor/execution/ExecutionHandler.java
@@ -13,17 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.service.connection.logicaldatabase.core.executor.execution.model;
+package com.oceanbase.odc.service.connection.logicaldatabase.core.executor.execution;
 
-public enum ExecutionStatus {
-    SUCCESS,
-    FAILED,
-    SKIPPED,
-    RUNNING,
-    PENDING,
-    TERMINATED;
+import java.sql.SQLException;
 
-    public boolean isCompleted() {
-        return this == SUCCESS || this == FAILED || this == SKIPPED || this == TERMINATED;
-    }
+public interface ExecutionHandler<Input, Result> {
+    ExecutionResult<Result> beforeExecute(ExecutionGroupContext<Input, Result> context);
+
+    ExecutionResult<Result> execute(ExecutionGroupContext<Input, Result> context) throws SQLException;
+
+    void terminate(ExecutionGroupContext<Input, Result> context) throws Exception;
 }
