+++ server/odc-service/src/main/java/com/oceanbase/odc/service/connection/logicaldatabase/model/SqlExecutionUnitResp.java
@@ -18,7 +18,7 @@ package com.oceanbase.odc.service.connection.logicaldatabase.model;
 import java.util.List;
 
 import com.oceanbase.odc.service.connection.database.model.Database;
-import com.oceanbase.odc.service.connection.logicaldatabase.core.executor.execution.model.ExecutionStatus;
+import com.oceanbase.odc.service.connection.logicaldatabase.core.executor.execution.ExecutionStatus;
 import com.oceanbase.odc.service.connection.logicaldatabase.core.executor.sql.SqlExecutionResultWrapper;
 import com.oceanbase.odc.service.connection.model.ConnectionConfig;
